Adding plugins settings

+++ oh-my-zsh/setup.sh
@@ -1,7 +1,5 @@
 #!/usr/sh
 
-ln -s $HOME/src/github.com/tatsuyano/dotfiles/oh-my-zsh/main.zsh $HOME/.oh-my-zsh/custom/main.zsh
-ln -s $HOME/src/github.com/tatsuyano/dotfiles/oh-my-zsh/peco.zsh $HOME/.oh-my-zsh/custom/peco.zsh
-
-ln -s $HOME/src/github.com/rupa/z/z.sh $HOME/.oh-my-zsh/custom/plugins/z.plugin.zsh
-ln -s $HOME/src/github.com/zsh-users/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh $HOME/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting.plugin.zsh
+[[ ! -e $HOME/.oh-my-zsh/custom/main.zsh ]] && ln -s $HOME/src/github.com/tatsuyano/dotfiles/oh-my-zsh/main.zsh $HOME/.oh-my-zsh/custom/main.zsh
+[[ ! -e $HOME/.oh-my-zsh/custom/peco.zsh ]] && ln -s $HOME/src/github.com/tatsuyano/dotfiles/oh-my-zsh/peco.zsh $HOME/.oh-my-zsh/custom/peco.zsh
+[[ ! -e $HOME/.oh-my-zsh/custom/plugins  ]] && ln -s $HOME/src/github.com/tatsuyano/dotfiles/oh-my-zsh/plugins  $HOME/.oh-my-zsh/custom/plugins
